refactor(benchmarks): relocate workload.py to src/benchmarks/
Move the pure-Python SQL-template workload driver (WorkloadExecutor,
WorkloadFileLoader, extract_workload_template_metadata) out of the PBT-specific
src/tuner/benchmark/ package into src/benchmarks/, beside the sysbench/tpch
C-binary executors. It imports only src.{database,config,utils} — no tuning
dependencies — and is consumed by the shared eval engine and the tuners/ utils.

- git mv workload.py -> src/benchmarks/ (history preserved)
- repoint the 3 importers (orchestrator, main, tuners/utils/executors)
- update doc path references (3 files)

Part of ADR-006 tuners/ unification. Relocate-first; no behavior change.
