+++ docs/guides/adding-workloads.md
@@ -19,7 +19,7 @@ For the architecture of how the workload is actually executed, see [workload-orc
 | **Tune PostgreSQL for your own application's queries** | **Custom JSON/YAML workload** (this guide) |
 | Mix custom and built-in queries | A custom workload — there is no first-class blending mechanism with sysbench/tpch |
 
-The custom workload runs through the [`WorkloadExecutor`](../../src/tuner/benchmark/workload.py), which is pure Python over psycopg2. It's slower than the C-binary sysbench/tpch executors because of GIL and round-trip overhead, but for moderately-complex queries the database execution time dominates the Python overhead and the measurement is still meaningful.
+The custom workload runs through the [`WorkloadExecutor`](../../src/benchmarks/workload.py), which is pure Python over psycopg2. It's slower than the C-binary sysbench/tpch executors because of GIL and round-trip overhead, but for moderately-complex queries the database execution time dominates the Python overhead and the measurement is still meaningful.
 
 ---
 
@@ -208,7 +208,7 @@ Before committing to a long PBT run:
 ```bash
 # Smoke-test the workload file parses
 python -c "
-from src.tuner.benchmark.workload import WorkloadFileLoader
+from src.benchmarks.workload import WorkloadFileLoader
 exe = WorkloadFileLoader.load_from_file('workloads/my_app.json')
 print(f'queries: {len(exe.queries)}')
 print(f'weights: {exe.weights}')
@@ -218,7 +218,7 @@ print(f'tables: {exe.num_tables}, size: {exe.table_size}')
 # Confirm the feature extractor parses the SQL templates correctly
 python -c "
 from src.utils.scoring.workload_features import WorkloadFeatureExtractor
-from src.tuner.benchmark.workload import extract_workload_template_metadata
+from src.benchmarks.workload import extract_workload_template_metadata
 meta = extract_workload_template_metadata('workloads/my_app.json')
 features = WorkloadFeatureExtractor.from_template_metadata(meta).extract()
 print(features)
