+++ src/tuners/utils/executors.py
@@ -17,7 +17,7 @@ from typing import Any, Dict, Optional
 
 from src.benchmarks.sysbench.executor import SysbenchExecutor
 from src.benchmarks.tpch.executor import TPCHExecutor
-from src.tuner.benchmark.workload import (
+from src.benchmarks.workload import (
     WorkloadFileLoader, extract_workload_template_metadata
 )
 from src.utils.metrics import WorkloadType
